% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Meta_matrix.R
\name{Meta_matrix}
\alias{Meta_matrix}
\title{Creat a ligand expression matrix}
\usage{
Meta_matrix(ave_expr, All_database,Synthetic_step2,Transporter,precusor_transporter)
}
\arguments{
\item{infile}{"./raw"}
}
\value{
A matrix of the infile
}
\description{
This function loads gene expression matrix and calulate ligand expression
matrix based on gene expression. It assume that each columns are variables
and column names are variable names. It assumes that rownames are gene symbols.
}
